requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
idna==3.11
python-dotenv==1.2.1
redis==7.1.0
requests==2.32.5
spotipy==2.25.2
urllib3==2.6.2
